requirements.txt
openai>=1.0.0
requests>=2.25.1
gitpython>=3.1.0
PyGithub>=1.58
python-dotenv>=0.21.0
lxml>=4.9.0